[readme] - Readme Updated
Updated the readme to include an explanation on how to use the Analytics
Admin API to pull GA4 account information. This includes an explanation
on how to enable the API in a Google Cloud Project and connect that
project to the script.

Also added an explanation on what the current GA4 utilities do.

Change-Id: I3d611eb1a595000f78e4690119c49f3d6caacbd3