[modify] - Modify most GA4 entities
This change mainly adds the ability to modify most GA4 entities under
modifyGA4Entities.js. This replaces the code in
modifyGACustomDefinitions.js and makes it applicable to nearly all GA4
entities. Each of the other GA4 files were slightly modified to
standardize the output of the data listed to a sheet. The functions in
analyticsAdminApi.js were simplified to have a single list function
for nearly all GA4 entities, a single create function for GA4, a single
delete function, and a single archive function. The individual list,
create, and archive functions were deleted.

The README file was updated to describe these new features.

Change-Id: Iba3da722b924351e60799e6915ce9f9c15170dae