[metrics] - Retrieve Metric Data
Added a feature where metric data can be requested in bulk from selected
views. This is similar to the "list events" features. A user selects
metrics in the header of the UA Metrics Request sheet and then the
script requests the data for each of those metrics from each selected
view. Each view's data is then saved as a separate row in the UA Metrics
Request sheet.

Change-Id: Ided10c98964677ec02a1e81e092c3e5ab23131ab